+++ CHANGELOG.md
@@ -14,8 +14,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   an extra path with modules.
 - `tt play`: support connection to a target instance by `application` name
   or `application:instance` name.
+- `tt coredump pack`: add several options to customize dump processing:
+  * `-e (--executable)`: specify Tarantool executable path.
+  * `-p (--pid)`: specify PID of the dumped process.
+  * `-t (--time)`: specify time of dump (seconds since the Epoch).
 
 ### Changed
+
+- `tt coredump pack`: by default tarantool executable path is obtained from
+  `PATH` instead of using the hardcoded path `/usr/bin/tarantool`.
 
 ### Fixed